# docs/swagger/swagger.json
{
  "basePath": "/",
  "definitions": {
...
    "V1CommerceCreateOrderResponse": {
      "properties": {
        "order_id": {
          "format": "uuid",
          "type": "string"
        },
        "order_status": {
          "$ref": "#/definitions/V1CommerceOrderStatus"
        },
        "replay": {
          "type": "boolean"
        },
        "vend_session_id": {
          "format": "uuid",
          "type": "string"
        },
        "vend_state": {
          "$ref": "#/definitions/V1VendSessionState"
        }
      },
      "required": [
        "order_id",
        "vend_session_id",
        "replay",
        "order_status",
        "vend_state"
      ],
      "type": "object"
    },
...
  },
  "host": "localhost:8080",
  "info": {
    "contact": {
      "name": "AVF Engineering",
      "url": "https://github.com/avf/avf-vending-system"
    },
    "description": "HTTP API for the AVF vending platform (cmd/api). All `/v1/*` routes require `Authorization: Bearer <JWT>` unless noted (health, optional metrics, Swagger when enabled). Bearer middleware failures use minimal JSON `{\"error\":{\"message\":\"...\"}}` without `error.code`. Most `/v1` handlers use `{\"error\":{\"code\":\"...\",\"message\":\"...\"}}` (see writeAPIError). **501** list stubs return `not_implemented` plus `capability` and `implemented:false`. **503** `capability_not_configured` is used when optional wiring (MQTT dispatch, commerce persistence, outbox defaults) is missing. Request tracing: responses echo `X-Request-ID` and `X-Correlation-ID` when middleware is enabled.",
    "license": {
      "name": "License terms apply per your deployment agreement; see repository NOTICE if present."
    },
    "termsOfService": "https://github.com/avf/avf-vending-system/tree/main/avf-vending-api",
    "title": "AVF Vending HTTP API",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/commerce/orders": {
      "post": {
        "consumes": [
          "json"
        ],
        "description": "Commerce routes not mounted when commerce service nil. Requires org on JWT for non-platform users. **Idempotency-Key** or **X-Idempotency-Key** required. Not configured \u2192 **503** `capability_not_configured` (`v1.commerce.persistence`). Rate limit \u2192 **429**.",
        "parameters": [
          {
            "description": "machine_id, product_id, slot_index, currency, subtotal_minor, tax_minor, total_minor",
            "in": "body",
            "name": "body",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "201": {
            "description": "Create order and initial vend session",
            "schema": {
              "$ref": "#/definitions/V1CommerceCreateOrderResponse"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "429": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "503": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1CapabilityNotConfiguredError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create order and initial vend session",
        "tags": [
          "Commerce"
        ]
      }
    },
    "/v1/commerce/orders/{orderId}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Optional **slot_index** query (default 0). Response nests `order`, `vend`, and `payment` (nullable).",
        "parameters": [
          {
            "description": "Order UUID",
            "in": "path",
            "name": "orderId",
            "required": true,
            "type": "string"
          },
          {
            "description": "Slot index (default 0)",
            "in": "query",
            "name": "slot_index",
            "required": false,
            "type": "integer"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "Checkout status for order",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "404": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "503": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1CapabilityNotConfiguredError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Checkout status for order",
        "tags": [
          "Commerce"
        ]
      }
    },
    "/v1/commerce/orders/{orderId}/payment-session": {
      "post": {
        "consumes": [
          "json"
        ],
        "description": "When commerce outbox env (topic/event/aggregate) is unset \u2192 **503** `capability_not_configured` (`v1.commerce.payment_session.outbox`). Requires idempotency header.",
        "parameters": [
          {
            "description": "Order UUID",
            "in": "path",
            "name": "orderId",
            "required": true,
            "type": "string"
          },
          {
            "description": "provider, payment_state, amount_minor, currency, outbox_payload_json",
            "in": "body",
            "name": "body",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "Start payment with outbox row",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "404": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "429": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "503": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1CapabilityNotConfiguredError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Start payment with outbox row",
        "tags": [
          "Commerce"
        ]
      }
    },
    "/v1/commerce/orders/{orderId}/payments/{paymentId}/webhooks": {
      "post": {
        "consumes": [
          "json"
        ],
        "parameters": [
          {
            "description": "Order UUID",
            "in": "path",
            "name": "orderId",
            "required": true,
            "type": "string"
          },
          {
            "description": "Payment UUID",
            "in": "path",
            "name": "paymentId",
            "required": true,
            "type": "string"
          },
          {
            "description": "provider, provider_reference, event_type, normalized_payment_state, payload_json, ...",
            "in": "body",
            "name": "body",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "Apply provider webhook",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "404": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "409": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "429": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "503": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1CapabilityNotConfiguredError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Apply provider webhook",
        "tags": [
          "Commerce"
        ]
      }
    },
    "/v1/commerce/orders/{orderId}/reconciliation": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Returns `{kind, status}` where **status** matches checkout status JSON.",
        "parameters": [
          {
            "description": "Order UUID",
            "in": "path",
            "name": "orderId",
            "required": true,
            "type": "string"
          },
          {
            "description": "Slot index",
            "in": "query",
            "name": "slot_index",
            "required": false,
            "type": "integer"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "Reconciliation snapshot wrapper",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "404": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "503": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1CapabilityNotConfiguredError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Reconciliation snapshot wrapper",
        "tags": [
          "Commerce"
        ]
      }
    },
    "/v1/commerce/orders/{orderId}/vend/failure": {
      "post": {
        "consumes": [
          "json"
        ],
        "parameters": [
          {
            "description": "Order UUID",
            "in": "path",
            "name": "orderId",
            "required": true,
            "type": "string"
          },
          {
            "description": "slot_index, failure_reason",
            "in": "body",
            "name": "body",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "Finalize vend failure",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "404": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "409": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "429": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "503": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1CapabilityNotConfiguredError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Finalize vend failure",
        "tags": [
          "Commerce"
        ]
      }
    },
    "/v1/commerce/orders/{orderId}/vend/start": {
      "post": {
        "consumes": [
          "json"
        ],
        "description": "Idempotency header required for write contract; value not yet used for dedupe on this transition. Illegal state \u2192 **409** `illegal_transition`.",
        "parameters": [
          {
            "description": "Order UUID",
            "in": "path",
            "name": "orderId",
            "required": true,
            "type": "string"
          },
          {
            "description": "slot_index",
            "in": "body",
            "name": "body",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "Advance vend to in_progress",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "404": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "409": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "429": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "503": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1CapabilityNotConfiguredError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Advance vend to in_progress",
        "tags": [
          "Commerce"
        ]
      }
    },
...
    "/v1/orders": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "**501** with capability `v1.orders.org_list`. Same auth and 403 behavior as payments list.",
        "produces": [
          "json"
        ],
        "responses": {
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "501": {
            "description": "List orders for organization",
            "schema": {
              "$ref": "#/definitions/V1NotImplementedError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List orders for organization",
        "tags": [
          "Commerce"
        ]
      }
    },
...
  },
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Send: Bearer <JWT>. JWT mode is selected by HTTP_AUTH_MODE (hs256 default, rs256_pem, rs256_jwks). 401/403 responses from this layer use JSON {\"error\":{\"message\":\"...\"}} without error.code.",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0",
...
}